refactor: replace launch link models into launch.dart
