+++ packages/spacex_api/lib/src/models/models.dart
@@ -1,7 +1,6 @@
 export 'company.dart';
 export 'filter.dart';
 export 'launch.dart';
-export 'launch_links.dart';
 export 'page.dart';
 export 'pagination_options.dart';
 export 'rocket.dart';
